Fixes https://github.com/grpc/grpc-dotnet/issues/369