@davidfowl 2019-07-10
Needs perf numbers