@davidfowl 2019-07-10
How much concurrency did you run with? The big concern with the global pool is contention.