@JamesNK 2019-07-10
The microbenchmark is single threaded.

The RPS benchmark has 16 workers hitting the server.