@davidfowl 2019-07-14
We should do a perf run