@JamesNK 2019-07-30
Will re-evaluate this in 5.0 when there is a better caching solution in ASP.NET Core.